Self employed remaining

+++ app/models/self_employed_account.rb
@@ -1,5 +1,5 @@
 class SelfEmployedAccount < Account
 
-  field :hourly_rate, type: Float, default: 0.0
+  field :hourly_rate, type: BigDecimal, default: 0.0
 
 end
